categories looping array pt1
